requirements.txt
asgiref==3.8.1
certifi==2024.8.30
click==8.1.7
diff-match-patch==20230430
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
rq==1.16.2
sqlparse==0.5.1
tablib==3.5.0
